requirements.txt
asgiref==3.5.0
Babel==2.9.1
black==22.3.0
certifi==2021.10.8
click==8.1.0
colorama==0.4.4
Django==4.0.3
django-countries==7.3.2
django-credit-cards==0.4.1
django-dbbackup==3.3.0
django-filter==21.1
django-money==2.1.1
djangorestframework==3.13.1
Markdown==3.3.6
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
py-moneyed==1.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2022.1
